Jivopis/requirements.txt
alabaster==0.7.12
anyascii==0.3.0
appdirs==1.4.4
argcomplete==1.10.0
arrow==0.13.1
asgiref==3.4.1
asn1crypto==0.24.0
asprise-ocr-sdk-python-api==15.3.0.8
astroid==2.5.3
asyncio==3.4.3
attrs==21.2.0
Automat==20.2.0
Babel==2.9.0
backports.functools-lru-cache==1.5
beautifulsoup4==4.8.0
bottle==0.12.16
Brotli==1.0.9
cattrs==1.8.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
charset-normalizer==2.0.7
cheroot==6.5.5
CherryPy==18.1.1
click==8.0.3
colorama==0.4.1
constantly==15.1.0
cryptography==2.2.2
cssselect==1.1.0
cursor==1.3.4
cycler==0.10.0
dash==0.41.0
dash-core-components==0.46.0
dash-html-components==0.15.0
dash-renderer==0.22.0
dash-table==3.6.0
decorator==4.4.0
Django==3.2.9
django-admin-list-filter-dropdown==1.0.3
django-admin-rangefilter==0.8.1
django-filter==21.1
django-modelcluster==5.2
django-taggit==1.5.1
django-tinymce==3.3.0
django-treebeard==4.5.1
djangorestframework==3.12.4
docutils==0.14
docx2txt==0.8
draftjs-exporter==2.1.7
EasyTkinter==1.1.0
EbookLib==0.17.1
et-xmlfile==1.0.1
extract-msg==0.23.1
feedparser==5.2.1
flake8==3.9.1
Flask==1.1.0
Flask-Admin==1.5.3
Flask-Compress==1.4.0
Flask-SeaSurf==0.2.2
Flask-SQLAlchemy==2.3.2
gitdb==4.0.9
greenlet==1.1.0
h2==3.2.0
hashlib-additional==1.0.1
hpack==3.0.0
html5lib==1.0.1
hupper==1.8.1
hyperframe==5.2.0
hyperlink==18.0.0
idna==2.6
imagesize==1.2.0
IMAPClient==2.1.0
importlib-metadata==4.0.1
importlib-resources==5.2.0
incremental==21.3.0
ipython-genutils==0.2.0
isort==5.8.0
itemadapter==0.3.0
itemloaders==1.0.4
itsdangerous==1.1.0
jaraco.classes==3.2.1
jaraco.collections==3.3.0
jaraco.functools==2.0
jaraco.text==3.5.1
jdcal==1.4
Jinja2==2.10.1
jmespath==0.10.0
joblib==1.1.0
jsonschema==3.0.1
jupyter-core==4.4.0
jwt==0.6.1
Kivy==2.0.0
kivy-deps.angle==0.3.0
kivy-deps.glew==0.3.0
kivy-deps.sdl2==0.3.1
Kivy-Garden==0.1.4
kiwisolver==1.0.1
l18n==2021.3
lazy-object-proxy==1.6.0
lxml==4.6.3
MarkupSafe==1.1.0
matplotlib==3.0.1
mccabe==0.6.1
more-itertools==7.0.0
multipledispatch==0.6.0
multitasking==0.0.9
mypy==0.812
mypy-extensions==0.4.3
mypy-lang==0.5.0
nbformat==4.4.0
numpy==1.21.4
olefile==0.46
opencv-python==4.1.0.25
openpyxl==2.6.2
packaging==20.9
panda==0.3.1
pandas==0.25.3
pandas-datareaders-unofficial==0.0.4
parsel==1.5.2
PasteDeploy==2.0.1
patsy==0.5.2
pdfminer.six==20181108
Pillow==6.0.0
plaster==1.0
plaster-pastedeploy==0.7
playsound==1.2.2
plotext==4.0.0
plotly==4.5.4
plotly-express==0.4.1
portend==2.5
priority==1.3.0
Protego==0.1.16
pyasn1==0.4.3
pyasn1-modules==0.2.1
pycodestyle==2.7.0
pyconfig==3.2.3
pycparser==2.18
pycryptodome==3.9.1
PyDispatcher==2.0.5
pyflakes==2.3.1
pygame==1.9.6
Pygments==2.3.1
PyHamcrest==1.9.0
pylint==2.7.4
pynput==1.6.8
pyOpenSSL==18.0.0
pyparsing==2.3.0
pypiwin32==223
PyQt5==5.11.3
PyQt5-stubs==5.11.3.3
PyQt5_sip==4.19.13
pyramid==1.10.4
pyrsistent==0.14.11
PySide2==5.12.2
pytesseract==0.3.0
python-dateutil==2.7.5
python-pptx==0.6.18
pytool==3.15.0
pytz==2021.3
pywin32==224
PyYAML==5.4.1
queuelib==1.5.0
requests==2.22.0
requests-cache==0.5.2
retrying==1.3.3
scipy==1.2.0
Scrapy==1.7.3
selenium==3.12.0
send-email==201212
service-identity==17.0.0
shiboken2==5.12.2
simplejson==3.16.0
simpy==3.0.11
six==1.16.0
sixel==0.0.7
smmap==5.0.0
snowballstemmer==2.1.0
sortedcontainers==2.1.0
soupsieve==1.9.4
SpeechRecognition==3.8.1
Sphinx==3.5.4
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
SQLAlchemy==1.2.13
sqlalchemy-s3sqlite==0.1
sqlparse==0.4.1
statsmodels==0.11.1
tablib==3.1.0
telepath==0.2
tempora==1.14.1
tenacity==8.0.1
termcolor==1.1.0
termgraph==0.5.3
terminalplot==0.3.0
termplotlib==0.3.9
textract==1.6.3
texttable==1.2.1
Theano==1.0.4
threadpoolctl==3.0.0
toml==0.10.2
tornado==6.0.2
tqdm==4.62.3
traitlets==4.3.2
translationstring==1.3
turtle==0.0.1
Twisted==19.7.0
twisted-iocpsupport==1.0.2
typed-ast==1.4.3
typing-extensions==3.7.4.3
tzlocal==1.5.1
ujson==4.0.2
url-normalize==1.4.3
urllib3==1.22
venusian==1.2.0
w3lib==1.21.0
wagtail==2.15.1
webencodings==0.5.1
WebOb==1.8.5
Werkzeug==0.15.4
Willow==1.4
win32-setctime==1.0.3
wrapt==1.12.1
WTForms==2.2.1
xlrd==1.2.0
XlsxWriter==3.0.2
xlwt==1.3.0
yaspin==2.1.0
yfinance==0.1.54
zc.lockfile==1.4
zipp==3.4.1
zope.deprecation==4.4.0
zope.interface==5.4.0
